+++ bpdm-gate/src/test/kotlin/org/eclipse/tractusx/bpdm/gate/auth/AuthInputManagerIT.kt
@@ -77,6 +77,10 @@ class AuthInputManagerIT @Autowired constructor(
         ),
         relationSharingState = RelationSharingStateExpectations(
             get = AuthExpectationType.Authorized
+        ),
+        relationChangelog = RelationChangelogExpectations(
+            getInput = AuthExpectationType.Authorized,
+            getOutput = AuthExpectationType.Forbidden
         )
     )
 ) {