+++ CHANGELOG.md
@@ -39,6 +39,7 @@ For changes to the BPDM Helm charts please consult the [changelog](charts/bpdm/C
 - BPDM Gate: Increased api version to v7 with simultaneous support for version v6([#1294](https://github.com/eclipse-tractusx/bpdm/issues/1294))
 - BPDM Gate: Add sharing state type and updatedAfter query filters to sharing state GET endpoint ([#1249](https://github.com/eclipse-tractusx/sig-release/issues/1249))
 - BPDM Pool: Add identifier type format and categories ([1297](https://github.com/eclipse-tractusx/bpdm/issues/1297))
+- BPDM Gate: Add changelog endpoints for business partner relations ([1328](https://github.com/eclipse-tractusx/bpdm/issues/1328))
 
 ### Changed